ad-token-server/requirements.txt
flask==3.0.3
flask-cors==4.0.0
azure-identity==1.17.1
requests==2.32.3